pyproject.toml
[project]
name = "jira-to-code-env"
version = "0.1.0"
description = "An environment where agents resolve Jira tickets by reading, writing, and testing code."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "pydantic",
    "fastapi",
    "uvicorn",
    "pytest",
    "openai",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Add "server" to the packages list
packages = ["src/jira_to_code", "server"]

[project.scripts]
# Update the script path
server = "server.app:main"